Decouple chunk generation from manager
