+++ Minecraft Clone/World/Chunks/ChunkManager.cs
@@ -13,6 +13,7 @@ public class ChunkManager
 {
     // needs no introduction
     public ChunkRenderer renderer = new ChunkRenderer();
+    public ChunkGenerator generator = new ChunkGenerator();
     public WorldGenerator worldGenerator = new WorldGenerator(512);
 
     public Vector3i currentChunkIndex = new();
@@ -31,7 +32,7 @@ public class ChunkManager
     ConcurrentDictionary<Vector3i, float> ExpiredChunkLifetimes = new ConcurrentDictionary<Vector3i, float>();
 
     // worker result classes, one for mesh worker and one for chunk generation worker
-    ConcurrentQueue<CompletedChunkBlocks> CompletedBlocksQueue = new();
+    ConcurrentQueue<ChunkGenerator.CompletedChunkBlocks> CompletedBlocksQueue = new();
     ConcurrentQueue<CompletedMesh> CompletedMeshQueue = new();
 
     // worker tasks list
@@ -45,76 +46,7 @@ public class ChunkManager
         maxChunkTasks = (int)(Environment.ProcessorCount -1);
     }
 
-    // chunk block generation delivery class
-    public class CompletedChunkBlocks
-    {
-        public Vector3i index;
-        public bool isEmpty;
-        public byte[] blocks;
-
-        public CompletedChunkBlocks(Vector3i index, byte[] blocks, bool isEmpty)
-        {
-            this.index = index;
-            this.blocks = blocks;
-            this.isEmpty = isEmpty;
-        }
-
-        public void Dispose()
-        {
-            Array.Clear(this.blocks);
-        }
-    }
-    // chunk's block generation kick-off fxn
-    public Task GenerationTask(Vector3i index)
-    {
-        ActiveChunks[index].SetState(ChunkState.GENERATING);
-        CancellationTokenSource cts = new CancellationTokenSource();
-        RunningTasksCTS.TryAdd(index, cts);
-
-        return Task.Run(async () =>
-        {
-
-            var result = await GenerateBlocks(index, cts.Token);
-            CompletedBlocksQueue.Enqueue(result);
-
-            //RunningTasks.TryRemove(index, out _);
-            //RunningTasksCTS.TryRemove(index, out _);
-        });
-    }
-    Task<CompletedChunkBlocks> GenerateBlocks(Vector3i chunkIndex, CancellationToken token)
-    {
-        var tempChunk = new Chunk();
-
-        return Task.Run(() =>
-        {
-            int totalBlocks = 0;
-
-            // for each block in the 16×16×16 volume...
-            for (int x = 0; x < Chunk.SIZE; x++)
-            {
-                for (int y = Chunk.SIZE - 1; y >= 0; y--)
-                {
-                    for (int z = 0; z < Chunk.SIZE; z++)
-                    {
-                        token.ThrowIfCancellationRequested();
-
-                        // compute world-space coordinate of this block
-                        int worldX = chunkIndex.X * Chunk.SIZE + x;
-                        int worldY = chunkIndex.Y * Chunk.SIZE + y;
-                        int worldZ = chunkIndex.Z * Chunk.SIZE + z;
-
-                        // offload world gen code to the generator. facade pattern
-                        BlockType type = worldGenerator.GetBlockAtWorldPos((worldX, worldY, worldZ));
-
-                        tempChunk.SetBlock(x, y, z, type);
-                        if (type != BlockType.AIR) totalBlocks++;
-                    }
-                }
-            }
-            return new CompletedChunkBlocks(chunkIndex, tempChunk.blocks, tempChunk.IsEmpty);
-        });
-    }
-
+    
     // chunk mesh delivery class
     public class CompletedMesh
     {
@@ -406,7 +338,12 @@ public class ChunkManager
                 {
                     case ChunkState.BIRTH:
                         if (RunningTasks.Count < maxChunkTasks)
-                            RunningTasks.TryAdd(idx, GenerationTask(idx));
+                        {
+                            ActiveChunks[idx].SetState(ChunkState.GENERATING);
+                            CancellationTokenSource cts = new CancellationTokenSource();
+                            RunningTasksCTS.TryAdd(idx, cts);
+                            RunningTasks.TryAdd(idx, generator.GenerationTask(idx, cts, worldGenerator, CompletedBlocksQueue));
+                        }
 
                         break;
                     case ChunkState.GENERATED:
